moved dynamic settings file, fixes and UX in plugin init / config, UX on plugin classes loading, new features / overhaul of auto-installer
moved dynamic settings file, fixes and UX in plugin init / config, UX on plugin classes loading, new features / overhaul of auto-installer
