+++ DOCUMENTATION-ETC/TODO.txt
@@ -150,6 +150,9 @@ DEBUGGING / TESTS / UX / QA / OTHER
 ############ NEXT DEBUG / TEST / QA / OTHER SCHEDULED WORK SESSION ###################################
 
 
+Test all the new FOLIO-INSTALL.bash features BEFORE v6.00.18 release, VIA A FRESH RASPBERRY PI OS INSTALLATION.
+
+
 Run testing on the overhauled plugin configuration / initiation system
 
 
@@ -172,6 +175,9 @@ Remove all linebreaks in javascript alert / confirm dialogue boxes, except for n
 
 
 ######################################################################################################
+
+
+Remove all forced line breaks in FOLIO-INSTALL.bash dialogue
 
 
 New FAQ entry "my text alerts aren't working, or are delayed"
