+++ DOCUMENTATION-ETC/changelog.txt
@@ -6044,6 +6044,16 @@ v6.00.2 AND HIGHER RESETS LIGHT (TIME PERIOD) CHARTS FROM ARCHIVAL DATA, ONE-TIM
 
 -BUG FIX: Fixed issues within the plugin configuration / initiation system
 
+-Debian-based auto-installer (FOLIO-INSTALL.bash) now runs auto-setup for enhanced Apache security (disables apache from showing product name / version number)
+
+-Debian-based auto-installer (FOLIO-INSTALL.bash) now has multiple OPTIONAL security-hardening automation settings (lets you pick the ones you want)
+
+-Debian-based auto-installer (FOLIO-INSTALL.bash) upgraded to auto-setup Apache PHP-FPM to run as the app user (for better security / compatibility)
+
+-Optimized plugin system user experience (error checks / correcting on primary configs, classes accessible from other plugins)
+
+-"dynamic-config-only.php" added to the main directory (now houses the dynamic configs used for POW mining calculators, etc)
+
 -Added admin / docs interface templates system for plugins (see /DOCUMENTATION-ETC/PLUGINS-README.txt)
 
 -Code cleanup / housekeeping
