+++ dynamic-config-only.php
