+++ templates/interface/desktop/php/user/user-sections/portfolio.php
@@ -1383,24 +1383,15 @@ zingchart.bind('marketcap_chart', 'label_click', function(e){
   	$this_plug = $plugin_key;
   		
   		if ( file_exists($plugin_init) && $plug_conf[$this_plug]['ui_location'] == 'more_stats' ) {
-  		
       	?>
-        <fieldset class='subsection_fieldset'>
-        	<legend class='subsection_legend'> <b><?=$plug_conf[$this_plug]['ui_name']?></b> </legend>
+          <fieldset class='subsection_fieldset'>
+             	<legend class='subsection_legend'> <b><?=$plug_conf[$this_plug]['ui_name']?></b> </legend>
       	<?php
-  	
-  			// This plugin's default class (only if the file exists)
-  			if ( file_exists($base_dir . '/plugins/'.$this_plug.'/plug-lib/plug-class.php') ) {
-  	        include($base_dir . '/plugins/'.$this_plug.'/plug-lib/plug-class.php');
-  			}
-  		
   		// This plugin's plug-init.php file (runs the plugin)
   		include($plugin_init);
-  		
-        ?>
-        </fieldset>
-        <?php
-    
+          ?>
+          </fieldset>
+          <?php
   		}
   		
   	// Reset $this_plug at end of loop
